Corrected Return Type in Subsystem
Int to double because it should return a double

+++ DeepSpace/src/main/cpp/subsystems/UltrasonicSensor.cpp
@@ -17,7 +17,7 @@ void UltrasonicSensor::InitDefaultCommand() {
   // SetDefaultCommand(new MySpecialCommand());
 }
 
-int UltrasonicSensor::GetDistanceToWall() {
+double UltrasonicSensor::GetDistanceToWall() {
   return m_ultrasonicSensor->GetRangeInches();
 
 }
